requirements.txt
absl-py==2.3.1
accelerate==1.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
airportsdata==20250523
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
blake3==1.0.5
blessed==1.20.0
cachetools==6.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
cloudpickle==3.1.1
cmake==4.0.3
compressed-tensors==0.9.4
contourpy==1.3.1
cuda-bindings==12.9.0
cuda-python==12.9.0
cupy-cuda12x==13.4.1
cycler==0.12.1
Deprecated==1.2.18
depyf==0.18.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
einops==0.8.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.17.0
flashinfer-python==0.2.7.post1
fonttools==4.57.0
frozenlist==1.6.2
fsspec==2025.2.0
fuzzywuzzy==0.18.0
gguf==0.17.0
googleapis-common-protos==1.70.0
gpustat==1.1.1
grpcio==1.72.1
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.2
idna==3.10
importlib_metadata==8.6.1
interegular==0.3.3
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lark==1.2.2
Levenshtein==0.27.1
llguidance==0.7.26
llvmlite==0.44.0
lm-format-enforcer==0.9.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mistral_common==1.5.6
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numba==0.61.2
numpy==1.26.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.3.9
nvidia-nvtx-cu12==12.6.77
openai==1.84.0
opencv-python-headless==4.11.0.86
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-semantic-conventions-ai==0.4.9
outlines==0.0.34
outlines_core==0.1.26
packaging==24.2
pandas==2.3.1
partial-json-parser==0.2.1.1.post5
pillow==11.2.1
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
propcache==0.3.1
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pycountry==24.6.1
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.12.2
ray==2.46.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.7
rouge==1.0.1
rouge_score==0.1.2
rpds-py==0.25.1
safetensors==0.5.3
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tiktoken==0.6.0
tokenizers==0.20.3
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.46.3
triton==3.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.3
uvloop==0.21.0
vllm_nccl_cu12==2.18.1.0.4.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
xformers==0.0.25
xgrammar==0.1.19
yarl==1.20.0
zipp==3.22.0
